fix(001): Fix Decimal JSON serialization in high_performers insight
Problem:
- generate_high_performers() was storing Decimal values in JSON field
- Python's round() on Decimal returns Decimal, not float
- SQLAlchemy JSON serialization failed with TypeError

Fix:
- Explicitly convert all numeric values to float before storing
- gain_loss_pct: float(round(value, 2))
- current_value: float(value)
- cost_basis: float(value)

Result:
- All 5 insights now generate successfully
- High performers insight displays correctly with gain/loss percentages

Changes:
- src/services/insight_generator.py:318-323

Testing:
✓ stocks-helper insight generate <portfolio-id>
✓ All 5 insight types generated successfully

+++ src/services/insight_generator.py
@@ -317,9 +317,9 @@ class InsightGenerator:
 
                     performers.append({
                         "ticker": holding.ticker,
-                        "gain_loss_pct": round(gain_loss_pct, 2),
-                        "current_value": current_value,
-                        "cost_basis": cost_basis,
+                        "gain_loss_pct": float(round(gain_loss_pct, 2)),
+                        "current_value": float(current_value),
+                        "cost_basis": float(cost_basis),
                     })
 
             # Sort by gain/loss % descending
